+++ SDTTool.py
@@ -136,7 +136,13 @@ def main(argv):
 	parser.add_argument('--hidedetails',  action='store_true', help='Hide the details of module classes and devices when printing documentation')
 	parser.add_argument('--markdowntables',  action='store_true', help='Format markdown output as tables for markdown')
 	parser.add_argument('--licensefile',  action='store', dest='licensefile', help='Add the text of license file to output files')
-	parser.add_argument('--domain',  action='store', dest='domain', help='Set the domain for the model')
+
+	oneM2MArgs = parser.add_argument_group('oneM2M')
+	oneM2MArgs.add_argument('--domain',  action='store', dest='domain', help='Set the domain for the model')
+	oneM2MArgs.add_argument('--withabbreviations',  action='store_true', dest='withabbreviations', help='Specify whether to generate additional output files with abbreviated identifiers.')
+	oneM2MArgs.add_argument('--abbreviationfile',  action='store', dest='abbreviationfile', help='Specify the file to store the list of abbreviations.')
+	oneM2MArgs.add_argument('--abbreviationlength',  action='store', dest='abbreviationlength', default='5', help='Specify the maximum length for abbreviations. The default is 5.')
+
 	requiredNamed = parser.add_argument_group('required arguments')
 	requiredNamed.add_argument('-i', '--infile', action='store', dest='inFile', required=True, help='The SDT input file to parse')
 	
@@ -155,6 +161,9 @@ def main(argv):
 	moreOptions['markdowntables'] = args.markdowntables
 	moreOptions['licensefile'] = args.licensefile
 	moreOptions['domain'] = args.domain
+	moreOptions['withabbreviations'] = args.withabbreviations
+	moreOptions['abbreviationfile'] = args.abbreviationfile
+	moreOptions['abbreviationlength'] = args.abbreviationlength
 
 
 	# Read input file. Check for correct format